backend/requirements.txt
#pip install -r backend/requirements.txt
django-grappelli>=4.0.2
# Ajouté pour améliorer l'interface admin
Django>=3.2,<4.0
celery>=5.2.0
django-celery-beat>=2.2.0
django-celery-results>=2.2.0
djangorestframework>=3.12.0
djangorestframework-simplejwt>=5.0.0
django-cors-headers==4.3.0
pillow>=8.0.0
psycopg2>=2.9.0
python-dotenv==1.0.0
redis>=4.0.0
weasyprint>=56.0
reportlab>=3.6.0